Raft Queue: Move from worker to core
The following is just a mechanical change, moving the queue from worker
to core. As rightly pointed out, importing workers from the apiserver
seems a little bit off.
