+++ worker/raft/apply.go
@@ -10,8 +10,8 @@ import (
 	"github.com/juju/clock"
 	"github.com/juju/errors"
 
+	"github.com/juju/juju/core/raft/queue"
 	"github.com/juju/juju/core/raftlease"
-	"github.com/juju/juju/worker/raft/queue"
 )
 
 const (
