+++ worker/raft/apply_test.go
@@ -16,7 +16,7 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/worker/raft/queue"
+	"github.com/juju/juju/core/raft/queue"
 )
 
 type applyOperationSuite struct {
