--- worker/raft/queue/package_test.go
@@ -1,14 +1,0 @@
-// Copyright 2021 Canonical Ltd.
-// Licensed under the AGPLv3, see LICENCE file for details.
-
-package queue
-
-import (
-	"testing"
-
-	gc "gopkg.in/check.v1"
-)
-
-func TestPackage(t *testing.T) {
-	gc.TestingT(t)
-}
